requirements.txt
absl-py
mock
numpy
rouge-score
sacrebleu
sentencepiece
tensorflow-text==1.15.0rc0
tensor2tensor==1.15.0
tensorflow-datasets==3.1.0
tensorflow-gpu==1.15.2
